refactor: remove rego scanner
