--- pkg/rego/runtime.go
@@ -1,28 +1,0 @@
-package rego
-
-import (
-	"os"
-	"strings"
-
-	"github.com/open-policy-agent/opa/ast"
-	"github.com/open-policy-agent/opa/version"
-)
-
-func addRuntimeValues() *ast.Term {
-	env := ast.NewObject()
-	for _, pair := range os.Environ() {
-		parts := strings.SplitN(pair, "=", 2)
-		if len(parts) == 1 {
-			env.Insert(ast.StringTerm(parts[0]), ast.NullTerm())
-		} else if len(parts) > 1 {
-			env.Insert(ast.StringTerm(parts[0]), ast.StringTerm(parts[1]))
-		}
-	}
-
-	obj := ast.NewObject()
-	obj.Insert(ast.StringTerm("env"), ast.NewTerm(env))
-	obj.Insert(ast.StringTerm("version"), ast.StringTerm(version.Version))
-	obj.Insert(ast.StringTerm("commit"), ast.StringTerm(version.Vcs))
-
-	return ast.NewTerm(obj)
-}
