--- pkg/rego/testdata/policies/invalid.rego
@@ -1,8 +1,0 @@
-# METADATA
-# schemas:
-# - input: schema["input"]
-package defsec.test_invalid
-
-deny {
-	input.Stages[0].Commands[0].FooBarNothingBurger == "lol"
-}
