--- pkg/rego/testdata/policies/valid.rego
@@ -1,8 +1,0 @@
-# METADATA
-# schemas:
-# - input: schema["input"]
-package defsec.test_valid
-
-deny {
-	input.Stages[0].Commands[0].Cmd == "lol"
-}
